+++ flutter_inappwebview_forge_android/android/src/main/kotlin/com/emirkanacar/flutter_inappwebview_forge_android/MyCookieManager.kt
@@ -415,6 +415,7 @@ open class MyCookieManager(initialPlugin: InAppWebViewFlutterPlugin) :
                 CookieSyncManager.createInstance(context).sync()
             }
         }
+        result.success(true)
     }
 
     override fun dispose() {